BracketAlgebras

Documentation for BracketAlgebras.

BracketAlgebras.Tabloid

BracketAlgebras.atomic_extensors

BracketAlgebras.standard_violation

BracketAlgebras.straighten

BracketAlgebras.Tabloid — Method

tabloid(b::BracketAlgebraElem, ordering::Vector{Int}=collect(1:parent(b).n))

Return the tabloid corresponding to the bracket monomial b ordered by ordering. ordering determines the order in which

Example: b = [1,2,3][3,4,5], ordering = [4,3,1,2,5] result: [4 3 5; 3 1 2]

BracketAlgebras.atomic_extensors — Method

atomic_extensors(b::BracketAlgebraElem)

Find all atomic extensors of the bracket algebra elem b. These are the equivalence classes of the relation p1 ~ p2 iff substituting p1 = p2 in b results in the zero element of the bracket algebra.

BracketAlgebras.standard_violation — Method

standard_violation(t::Tabloid)

Return the index of the first violation to the standardness of t, i.e. the first index where t[i,j] > t[i+1, j] with regard to the ordering. Otherwise return nothing. Example: t = [1 2 3; 1 4 5; 1 5 6; 2 3 4] with ordering [1,2,3,4,5,6] => return (3,2) t = [1 2 3; 1 2 4] with ordering [1,2,3,4] => return nothing

BracketAlgebras.straighten — Method

straighten(b::BracketAlgebraElem, ordering::AbstractVector{<:Integer} = collect(1:parent(b).n); chek::Bool = true)

Perform the straightening algorithm to the BracketAlgebraElem b with monomial ordering induced by ordering[1] < ordering[2] < ...

For details see Sturmfels 2008, chapter 3.1, Handbook of Geometric Constraint System Principles, chapater 4.3